规则已经可以从客户端下发到驱动并添加进驱动链表进行拦截

+++ CFileManager.h
@@ -47,7 +47,7 @@ typedef struct filenames filenames;
 
 //发送数据结构体
 typedef struct Data {
-	int command;
+	unsigned long command;
 	wchar_t filename[MAX_PATH];
 
 }Data,*pData;
